% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R
\name{cause}
\alias{cause}
\alias{cause<-}
\title{Get or set the \code{"cause"} attribute}
\usage{
cause(x)

cause(x) <- value
}
\arguments{
\item{x}{Any variable.}

\item{value}{Passed to \code{gettextf} and stored in the \code{"cause"}
attribute.}
}
\value{
The get method returns the \code{"cause"} attribute.
}
\description{
Gets or sets the \code{"cause"} (of failure) attribute of a variable.
}
\examples{
# Scalar case
yn <- is_identical_to_true(FALSE)
cause(yn)

# Vector case
yn <- is_true(c(TRUE, FALSE, NA))
cause(yn)
}
\seealso{
\code{\link{set_cause}}
}
