% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R
\name{print.scalar_with_cause}
\alias{print.scalar_with_cause}
\alias{print.vector_with_cause}
\title{Print methods for objects with a cause attribute}
\usage{
\method{print}{scalar_with_cause}(x, ...)

\method{print}{vector_with_cause}(x, na_ignore = FALSE, n_to_show = 10, ...)
}
\arguments{
\item{x}{an object of class \code{scalar_with_cause} or
\code{vector_with_cause}.}

\item{...}{Currently unused.}

\item{na_ignore}{A logical value.  If \code{FALSE}, \code{NA} values
are printed; otherwise they do not.  Like \code{na.rm} in many
stats package functions, except that the position of the failing
values does not change.}

\item{n_to_show}{A natural number.  The maximum number of failures 
to show.}
}
\description{
Prints objects of class \code{scalar_with_cause} and 
\code{vector_with_cause}.
}
